+++ src/app/SubmitAction.php
@@ -21,6 +21,8 @@ class SubmitAction extends AbstractAction
 
             $reportFile->moveTo($fullPath);
             $dmarcFiles[] = $fullPath;
+
+            $this->fixEncodingIssues($fullPath);
         }
 
         /** @var \Solaris\DmarcAggregateParser $dmarcParser */
@@ -40,5 +42,34 @@ class SubmitAction extends AbstractAction
         }
 
         return $response->withJson(['status' => 'success']);
+    }
+
+    /**
+     * Determines whether given data is b64 encoded.
+     *
+     * @param $data
+     *
+     * @return bool
+     *
+     * @see http://stackoverflow.com/a/34982057/7362396
+     */
+    public static function is_base64_encoded($data)
+    {
+        return preg_match('%^[a-zA-Z0-9/+]*={0,2}$%', $data) === 1;
+    }
+
+    /**
+     * Checks for and fixes certain encoding issues that can happen to the reports in the mails.
+     *
+     * @param string $fullPath
+     */
+    public static function fixEncodingIssues(string $fullPath)
+    {
+        $content = file_get_contents($fullPath);
+        $contentNewlinesStripped = preg_replace('/\s+/', '', $content);
+
+        if (static::is_base64_encoded($contentNewlinesStripped)) {
+            file_put_contents($fullPath, base64_decode($contentNewlinesStripped));
+        }
     }
 }